# api/openapi.yaml
openapi: 3.0.0
info:
  title: cert-study-base API
  description: セキュリティ資格勉強会管理システムのREST API
  version: 0.1.0
servers:
  - url: http://localhost:8080
    description: ローカル開発環境
paths:
...
  /api/v1/auth/login:
    post:
      summary: ログイン
      description: username とパスワードで認証し、JWTトークンを返す。
      operationId: login
      tags:
        - auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: 認証成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/auth/logout:
    post:
      summary: ログアウト
      description: |
        ログアウトエンドポイント。
        サーバーサイドのトークン無効化リストは保持しない。
        クライアント側でトークンを削除することでセッションを終了する。
        このエンドポイントは将来的なサーバーサイド無効化への拡張口として提供する。
      operationId: logout
      tags:
        - auth
      security:
        - bearerAuth: []
      responses:
        "200":
          description: ログアウト成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      message:
                        type: string
                        example: ログアウトしました。クライアント側でトークンを削除してください。
                  error:
                    type: string
                    nullable: true
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/users:
    get:
      summary: ユーザー一覧取得（admin のみ）
      operationId: listUsers
      tags:
        - users
      security:
        - bearerAuth: []
      responses:
        "200":
          description: ユーザー一覧
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/User"
                  error:
                    type: string
                    nullable: true
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalServerError"

    post:
      summary: ユーザー作成（admin のみ）
      operationId: createUser
      tags:
        - users
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: ユーザー作成成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/User"
                  error:
                    type: string
                    nullable: true
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/users/{id}:
    get:
      summary: ユーザー取得（admin のみ）
      operationId: getUser
      tags:
        - users
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/UserID"
      responses:
        "200":
          description: ユーザー情報
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/User"
                  error:
                    type: string
                    nullable: true
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

    put:
      summary: ユーザー更新（admin のみ）
      operationId: updateUser
      tags:
        - users
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/UserID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: ユーザー更新成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/User"
                  error:
                    type: string
                    nullable: true
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalServerError"

    delete:
      summary: ユーザー削除（admin のみ）
      operationId: deleteUser
      tags:
        - users
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/UserID"
      responses:
        "200":
          description: ユーザー削除成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      message:
                        type: string
                        example: ユーザーを削除しました
                  error:
                    type: string
                    nullable: true
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/questions:
    post:
      summary: 問題作成（認証済みユーザー）
      description: Markdown形式の問題・解答・解説・議論点メモを作成します。デフォルトは status=draft、visibility_scope=all です。
      operationId: createQuestion
      tags:
        - questions
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQuestionRequest"
      responses:
        "201":
          description: 問題作成成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Question"
                  error:
                    type: string
                    nullable: true
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"

    get:
      summary: 問題一覧取得（認証済みユーザー）
      description: |
        登録されている問題の一覧を返します。
        クエリパラメータによるタグフィルタリング・キーワード検索・ページネーションに対応しています。
        - `tag_ids`: カンマ区切りのタグID一覧。複数指定時はAND絞り込みです。
        - `keyword`: タイトル・問題文・解説・メモを対象とするキーワード検索（部分一致）。
        - `page` / `per_page`: オフセット方式ページネーション。
      operationId: listQuestions
      tags:
        - questions
      security:
        - bearerAuth: []
      parameters:
        - name: tag_ids
          in: query
          required: false
          description: タグIDのカンマ区切り一覧（複数指定時はAND絞り込み）
          schema:
            type: string
            example: "tag-uuid-1,tag-uuid-2"
        - name: keyword
          in: query
          required: false
          description: キーワード検索（title / body / explanation / memo を対象とした部分一致）
          schema:
            type: string
            example: "暗号化"
        - name: page
          in: query
          required: false
          description: 取得するページ番号（1始まり。省略時は1）
          schema:
            type: integer
            minimum: 1
            default: 1
            example: 1
        - name: per_page
          in: query
          required: false
          description: 1ページあたりの件数（省略時は20、最大100）
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
            example: 20
      responses:
        "200":
          description: 問題一覧（ページネーション情報付き）
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/QuestionListResponse"
                  error:
                    type: string
                    nullable: true
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/questions/{id}:
    get:
      summary: 問題詳細取得（認証済みユーザー）
      description: 指定IDの問題を取得します。Markdownテキスト（body / answer / explanation / memo）を含みます。
      operationId: getQuestion
      tags:
        - questions
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/QuestionID"
      responses:
        "200":
          description: 問題詳細
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Question"
                  error:
                    type: string
                    nullable: true
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

    put:
      summary: 問題更新（作成者本人または admin のみ）
      description: 指定IDの問題を更新します。作成者本人または admin ロールのみ操作可能です。
      operationId: updateQuestion
      tags:
        - questions
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/QuestionID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateQuestionRequest"
      responses:
        "200":
          description: 問題更新成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Question"
                  error:
                    type: string
                    nullable: true
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

    delete:
      summary: 問題削除（作成者本人または admin のみ）
      description: 指定IDの問題を削除します。作成者本人または admin ロールのみ操作可能です。
      operationId: deleteQuestion
      tags:
        - questions
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/QuestionID"
      responses:
        "200":
          description: 問題削除成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      message:
                        type: string
                        example: 問題を削除しました
                  error:
                    type: string
                    nullable: true
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/questions/{id}/visibility:
    patch:
      summary: 問題の公開設定変更（作成者本人または admin のみ）
      description: |
        指定IDの問題の公開ステータス・公開範囲・公開対象チームIDを変更します。
        作成者本人または admin ロールのみ操作可能です。
        - status=draft / private の場合、作成者のみが閲覧できます。
        - status=published かつ visibility_scope=all の場合、全ログインユーザーが閲覧できます。
        - status=published かつ visibility_scope=team の場合、published_team_ids に指定したチームのメンバーのみ閲覧できます。
      operationId: updateQuestionVisibility
      tags:
        - questions
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/QuestionID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateQuestionVisibilityRequest"
      responses:
        "200":
          description: 公開設定更新成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Question"
                  error:
                    type: string
                    nullable: true
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/questions/{id}/comments:
    post:
      summary: コメント投稿（閲覧権限を持つ認証済みユーザー）
      description: |
        指定IDの問題にコメントを投稿します。
        その問題の閲覧権限を持つユーザーのみ投稿可能です。
        - status=published かつ visibility_scope=all → 全ログインユーザー
        - status=published かつ visibility_scope=team → 指定チームのメンバーのみ
        - status=draft / private → 作成者本人のみ（実質コメント不可）
      operationId: createComment
      tags:
        - comments
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/QuestionID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCommentRequest"
      responses:
        "201":
          description: コメント投稿成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Comment"
                  error:
                    type: string
                    nullable: true
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

    get:
      summary: コメント一覧取得（閲覧権限を持つ認証済みユーザー）
      description: |
        指定IDの問題のコメント一覧を投稿日時の昇順で返します。
        レスポンスには投稿者の display_name を含みます。
        その問題の閲覧権限を持つユーザーのみ取得可能です。
      operationId: listComments
      tags:
        - comments
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/QuestionID"
      responses:
        "200":
          description: コメント一覧
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Comment"
                  error:
                    type: string
                    nullable: true
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/questions/{id}/comments/{comment_id}:
    put:
      summary: コメント編集（投稿者本人のみ）
      description: |
        指定コメントを編集します。
        投稿者本人のみ編集可能です。その問題の閲覧権限も必要です。
      operationId: updateComment
      tags:
        - comments
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/QuestionID"
        - $ref: "#/components/parameters/CommentID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCommentRequest"
      responses:
        "200":
          description: コメント編集成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Comment"
                  error:
                    type: string
                    nullable: true
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

    delete:
      summary: コメント削除（投稿者本人または admin）
      description: |
        指定コメントを削除します。
        投稿者本人または admin ロールのみ削除可能です。その問題の閲覧権限も必要です。
      operationId: deleteComment
      tags:
        - comments
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/QuestionID"
        - $ref: "#/components/parameters/CommentID"
      responses:
        "200":
          description: コメント削除成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      message:
                        type: string
                        example: コメントを削除しました
                  error:
                    type: string
                    nullable: true
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/teams:
    post:
      summary: チーム作成（admin / teamowner のみ）
      operationId: createTeam
      tags:
        - teams
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTeamRequest"
      responses:
        "201":
          description: チーム作成成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Team"
                  error:
                    type: string
                    nullable: true
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalServerError"

    get:
      summary: チーム一覧取得（admin は全件、それ以外は自分が所属するチームのみ）
      operationId: listTeams
      tags:
        - teams
      security:
        - bearerAuth: []
      responses:
        "200":
          description: チーム一覧
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Team"
                  error:
                    type: string
                    nullable: true
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /api/v1/teams/{id}/members:
    post:
      summary: メンバー追加（owner_id 本人または admin のみ）
      operationId: addTeamMember
      tags:
        - teams
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/TeamID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddTeamMemberRequest"
      responses:
        "201":
          description: メンバー追加成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/TeamMember"
                  error:
                    type: string
                    nullable: true
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/teams/{id}/members/{user_id}:
    delete:
      summary: メンバー除外（per-team owner または admin のみ）
      operationId: removeTeamMember
      tags:
        - teams
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/TeamID"
        - name: user_id
          in: path
          required: true
          description: 除外するユーザーID（UUID形式）
          schema:
            type: string
            format: uuid
            example: "550e8400-e29b-41d4-a716-446655440001"
      responses:
        "200":
          description: メンバー除外成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      message:
                        type: string
                        example: メンバーを除外しました
                  error:
                    type: string
                    nullable: true
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    UserID:
      name: id
      in: path
      required: true
      description: ユーザーID（UUID形式）
      schema:
        type: string
        format: uuid
        example: "550e8400-e29b-41d4-a716-446655440000"
...
  schemas:
...
    LoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          description: ログインID
          example: "alice"
        password:
          type: string
          description: パスワード（平文）
          minLength: 8
          example: "password123"
    LoginResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            token:
              type: string
              description: JWTアクセストークン
              example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
            user:
              $ref: "#/components/schemas/User"
        error:
          type: string
          nullable: true
...
  responses:
    BadRequest:
      description: リクエストが不正
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: 認証が必要、またはトークンが無効
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: 権限がない
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: リソースが見つからない
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: リソースが競合（username・email の重複など）
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    InternalServerError:
      description: サーバー内部エラー
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"